Get rid of of PODArray copypaste. [#CLICKHOUSE-3084]
